Промежуточная фикация
